Switch mocha to jest
This commit replaces mocha with jest. Also this commit removes
dependency on babel-plugin-webpack-loaders plugin because of
the following reasons:
1. It seems that the author deprecated and abandoned this plugin
and he suggests not to use it and to use mocks and native test utilities
tools instead.
2. This plugin conflicts with webpack 2 and because of the reason one
it is unlikely that it will ever support it.

In order to remove dependency on the aforementioned plugin and
keep using css modules in the same time the following trick is used:
All components that need to be selected in the end-to-end test receive
data-tid (TestID) and data-tclass (TestCLASS) properties. It is necessary
because without running webpack we cannot use class names to select
components because we do not know what class names css loader
actually generates for us.